refactor(app): extract useActiveCampaignDebugLog from useAppModel
